Refactoring and restructuring the project files.
