+++ ApiApplication/Domain/IImdbStatusHostedService.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Extensions.Hosting;
 using System;
 
-namespace ApiApplication.Services
+namespace ApiApplication.Domain
 {
     public interface IImdbStatusHostedService : IHostedService, IDisposable
     {
