+++ ApiApplication/Domain/ImdbStatusHostedService.cs
@@ -4,7 +4,7 @@ using System;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace ApiApplication.Services
+namespace ApiApplication.Domain
 {
     public class ImdbStatusHostedService : BackgroundService, IImdbStatusHostedService
     {
